requirements.txt
matplotlib
pydensecrf
numpy
Pillow
torch
torchvision
pyaml
